requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.1.1
idna==3.10
iniconfig==2.0.0
numpy==1.26.4
packaging==24.1
pandas==1.5.3
pip==24.0
pluggy==1.5.0
psycopg2==2.9.9
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
setuptools==65.5.0
six==1.16.0
SQLAlchemy==1.4.53
sqlalchemy-redshift==0.8.14
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
